// src/sdk/types/enums.ts
export enum WAConfigEnum {
    BaseURL = "WA_BASE_URL",
    AppId = "M4D_APP_ID",
    AppSecret = "M4D_APP_SECRET",
    PhoneNumberId = "WA_PHONE_NUMBER_ID",
    BusinessAcctId = "WA_BUSINESS_ACCOUNT_ID",
    APIVersion = "CLOUD_API_VERSION",
    AccessToken = "CLOUD_API_ACCESS_TOKEN",
    WebhookEndpoint = "WEBHOOK_ENDPOINT",
    WebhookVerificationToken = "WEBHOOK_VERIFICATION_TOKEN",
    ListenerPort = "LISTENER_PORT",
    MaxRetriesAfterWait = "MAX_RETRIES_AFTER_WAIT",
    RequestTimeout = "REQUEST_TIMEOUT",
    Debug = "DEBUG",
}

export enum WARequiredConfigEnum {
    APIVersion = "CLOUD_API_VERSION",
    AccessToken = "CLOUD_API_ACCESS_TOKEN",
    PhoneNumberId = "WA_PHONE_NUMBER_ID",
}